adagent/requirements.txt
langchain-google-genai
langchain
langchain-community
langchain-core
pypdf
chromadb
fastapi
uvicorn
python-dotenv
unstructured
langchain-agent
google-api-python-client
